fix double bug 102

+++ src/main/java/com/richo/test/findbugs/App.java
@@ -24,7 +24,7 @@ public class App
     public static void main( String[] args )
     {
         System.out.println( "Hello World!" );
-        int testresult;
+        double testresult;
         testresult=1+2.0;
         boolean bool = api.nullable().contains("asd"); //This should not warn for NPE
         System.out.println(bool);
